requirements.txt
anthropic>=0.45.0
feedparser>=6.0.0
firecrawl-py>=1.0.0
python-dotenv>=1.0.0
pytz>=2024.1
pyyaml>=6.0
requests>=2.31.0
pytest>=8.0.0
